refactor: delete ./frontend/app/src/main/java/com/example/pkucat/ui dir, move four directories in /ui above
