+++ src/frontend/app/src/main/java/com/example/pkucat/archive/ArchiveFragment.java
@@ -1,4 +1,4 @@
-package com.example.pkucat.ui.setting;
+package com.example.pkucat.archive;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -14,16 +14,16 @@ import androidx.lifecycle.ViewModelProviders;
 
 import com.example.pkucat.R;
 
-public class SettingFragment extends Fragment {
-    private SettingViewModel settingViewModel;
+public class ArchiveFragment extends Fragment {
+    private ArchiveViewModel archiveViewModel;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        settingViewModel =
-                ViewModelProviders.of(this).get(SettingViewModel.class);
-        View root = inflater.inflate(R.layout.fragment_setting, container, false);
-        final TextView textView = root.findViewById(R.id.text_setting);
-        settingViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
+        archiveViewModel =
+                ViewModelProviders.of(this).get(ArchiveViewModel.class);
+        View root = inflater.inflate(R.layout.fragment_archive, container, false);
+        final TextView textView = root.findViewById(R.id.text_archive);
+        archiveViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
             @Override
             public void onChanged(@Nullable String s) {
                 textView.setText(s);
