+++ src/frontend/app/src/main/java/com/example/pkucat/post/list/PostFragment.java
@@ -1,27 +1,23 @@
-package com.example.pkucat.ui.post.ui;
+package com.example.pkucat.post.list;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
 import androidx.recyclerview.widget.DividerItemDecoration;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.pkucat.R;
-import com.example.pkucat.ui.post.PostEntity;
+import com.example.pkucat.post.PostEntity;
 
-import java.net.ServerSocket;
 import java.util.ArrayList;
-import java.util.List;
 
 public class PostFragment extends Fragment {
