+++ src/frontend/app/src/main/java/com/example/pkucat/post/list/PostViewModel.java
@@ -1,4 +1,4 @@
-package com.example.pkucat.ui.post.ui;
+package com.example.pkucat.post.list;
 
 import androidx.arch.core.util.Function;
 import androidx.lifecycle.LiveData;
