+++ src/frontend/app/src/main/java/com/example/pkucat/post/list/RecyclerViewAdapter.java
@@ -1,10 +1,9 @@
-package com.example.pkucat.ui.post.ui;
+package com.example.pkucat.post.list;
 
 import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AdapterView;
 import android.widget.ImageView;
 import android.widget.TextView;
 
@@ -12,9 +11,10 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.pkucat.R;
-import com.example.pkucat.ui.post.PostEntity;
+import com.example.pkucat.post.PostEntity;
 
 import java.util.ArrayList;
+
 
 public class RecyclerViewAdapter extends RecyclerView.Adapter<RecyclerViewAdapter.MyViewHolder> {
