+++ src/frontend/app/src/main/java/com/example/pkucat/setting/SettingFragment.java
@@ -1,4 +1,4 @@
-package com.example.pkucat.ui.recognition;
+package com.example.pkucat.setting;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -14,16 +14,16 @@ import androidx.lifecycle.ViewModelProviders;
 
 import com.example.pkucat.R;
 
-public class RecognitionFragment extends Fragment {
-    private RecognitionViewModel recognitionViewModel;
+public class SettingFragment extends Fragment {
+    private SettingViewModel settingViewModel;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        recognitionViewModel =
-                ViewModelProviders.of(this).get(RecognitionViewModel.class);
-        View root = inflater.inflate(R.layout.fragment_recognition, container, false);
-        final TextView textView = root.findViewById(R.id.text_recognition);
-        recognitionViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
+        settingViewModel =
+                ViewModelProviders.of(this).get(SettingViewModel.class);
+        View root = inflater.inflate(R.layout.fragment_setting, container, false);
+        final TextView textView = root.findViewById(R.id.text_setting);
+        settingViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
             @Override
             public void onChanged(@Nullable String s) {
                 textView.setText(s);
