+++ src/frontend/app/src/main/res/layout/fragment_archive.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.archive.ArchiveFragment">
+    tools:context=".archive.ArchiveFragment">
 
     <TextView
         android:id="@+id/text_archive"
