+++ src/frontend/app/src/main/res/layout/fragment_post.xml
@@ -5,7 +5,7 @@
         android:id="@+id/constraintLayout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.post.ui.PostFragment">
+        tools:context=".post.list.PostFragment">
 
 <!---
         <TextView
