+++ src/frontend/app/src/main/res/layout/fragment_recognition.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.recognition.RecognitionFragment">
+    tools:context=".recognition.RecognitionFragment">
 
     <TextView
         android:id="@+id/text_recognition"
